Add Thermo Fisher thermocycler vendor module (ProFlex/ATC)
- ThermoFisherThermocyclerDriver: single class for all SCPI I/O, auth,
  block discovery, power, temperature control, protocol execution
- Three capability backends per block:
  - ThermoFisherBlockBackend (TemperatureControllerBackend)
  - ThermoFisherLidBackend (TemperatureControllerBackend)
  - ThermoFisherThermocyclingBackend (ThermocyclingBackend)
- ThermoFisherThermocycler device with factory classmethods:
  proflex_single_block (1 block/6 zones), proflex_three_block (3 blocks/2 zones),
  atc (1 block/3 zones + lid control)
- Each block gets its own ThermocyclingCapability — run 3 PCR protocols simultaneously
- Legacy ThermoFisherThermocyclerBackend gutted to thin adapter (~275 lines, was ~1050)
- All 7 legacy tests pass
