+++ pylabrobot/thermo_fisher/__init__.py
@@ -1,0 +1,7 @@
+from .thermocycler import (
+  ThermoFisherBlockBackend,
+  ThermoFisherLidBackend,
+  ThermoFisherThermocycler,
+  ThermoFisherThermocyclerDriver,
+  ThermoFisherThermocyclingBackend,
+)
